macos specific build hacks

+++ tools/build.sh
@@ -5,9 +5,11 @@ PARENT_DIR=$(dirname $SCRIPT_DIR)
 
 echo $PARENT_DIR
 cd $PARENT_DIR/espeak-ng
-rm CHANGELOG.md
-echo "Changelog dummy" > ChangeLog.md
-echo "No news" > NEWS
+if [[ "$(uname)" == "Darwin" ]]; then
+    brew install automake libtool autoconf
+    rm CHANGELOG.md
+    echo "Changelog dummy" > ChangeLog.md
+fi
 ls
 ./autogen.sh
 ./configure --without-klatt --without-speechplayer --without-mbrola --without-sonic --without-async
